Added file structure to house the 3 configurations of the application
